requirements.txt
#python 3.10.9 64
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.1
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
scipy==1.14.1
six==1.16.0
spectral==0.23.1
tqdm==4.67.1
gitpython